requirements.txt
opencv-python==4.8.1.78
numpy==1.26.2
mysql-connector-python==8.2.0
mysqlclient==2.2.0
imutils==0.5.4
pytesseract==0.3.10
